Removed test_case files from lib directory.
