+++ test/rdoc/support/formatter_test_case.rb
