+++ test/rdoc/support/text_formatter_test_case.rb
